% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{shot_zone_summary}
\alias{shot_zone_summary}
\title{Summarise shot success by year and zone}
\usage{
shot_zone_summary(data = nba_shots)
}
\arguments{
\item{data}{A tibble like \code{nba_shots}.}
}
\value{
A tibble with columns: \code{year}, \code{shot_zone_basic}, and \code{accuracy}.
}
\description{
Provides aggregated shot accuracy by year and basic zone classification.
}
\examples{
shot_zone_summary()
}
