% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_nbashots_app}
\alias{run_nbashots_app}
\title{Launch the NBA Shiny app}
\usage{
run_nbashots_app()
}
\value{
A running Shiny app.
}
\description{
Opens the interactive explorer bundled in this package.
}
